requirements.txt
pyqt5
qtawesome
loguru
pyautogui
pyyaml
opencv_python
numpy
pyinstaller